Test the counter method
