+++ src/counter.js
@@ -1,0 +1,4 @@
+const counter = (reciveCommentsApi, movie) => reciveCommentsApi(movie.id)
+  .then((element) => (element.json()));
+
+export default counter;
